requirements.txt
djangorestframework
django-cors-headers

pillow
djoser
pytils

celery
redis

django-cleanup
django-ckeditor
django-filter
pytils
segno




